+++ api.schedule/schedule-main-service/internal/model/lesson.go
@@ -3,14 +3,14 @@ package model
 import "time"
 
 type Lesson struct {
-	ID           uint64
-	Number       uint8
+	ID           int64
+	Number       int64
 	Start        time.Time
 	End          time.Time
 	Duration     time.Time
-	TimetableID  uint64
-	DisciplineID uint64
-	TeacherIDs   []uint64
-	ClassroomIDs []uint64
+	TimetableID  int64
+	DisciplineID int64
+	TeacherIDs   []int64
+	ClassroomIDs []int64
 	IsChanged    bool
 }
