+++ api.schedule/schedule-main-service/internal/model/term.go
@@ -1,7 +1,7 @@
 package model
 
 type Term struct {
-	ID                 uint8
+	ID                 int64
 	Course             uint8
 	CourseRelativeTerm uint8
 }
